requirements.txt
pyttsx3
speech_recognition
SpeechRecognition
wikipedia
pyautogui
pywhatkit
PyAudio
google-genai
cv2
requests
webbrowser